requirements.txt
Pillow==9.4.0
pyserial==3.5